test(db): move shared enqueue fixtures into conftest (#124)
